Restructure core functions to reduce unnecessary wrapping.
